### 2022-05-02 #4362 @editorialbot
[REVIEW]: Pose2Sim: An Open-Source Python Package for multiview markerless kinematics
**Submitting author:** @DavidPagnon (<a href="http://orcid.org/0000-0002-6891-8331">David Pagnon</a>)
**Repository:** https://github.com/perfanalytics/pose2sim
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
