### 2022-05-11 #4362 @davidpagnon
Dear reviewers,
I suppose it is not really advised to change the code once it's under review, but I wanted to make Poe2Sim available through pip. It made me slightly modify the file structure, although none of the core code has been changed. I tested it all, updated the Readme, and pushed the changes today.
I don't think it will make a big difference, but if you already cloned the repository, you should probably `git pull` it or clone it again to make sure you have the last version.
Best regards,
David
