### 2022-05-29 #4362 @CVHammond
see [this issue](https://github.com/davidpagnon/pose2sim/issues/2) about potential syntax issue
