### 2022-05-31 #4362 @lambdaloop
See https://github.com/davidpagnon/pose2sim/issues/3 about an issue for running code on Linux
