### 2022-05-31 #4362 @lambdaloop
See https://github.com/davidpagnon/pose2sim/issues/5 about specifying the version of OpenSim
