### 2022-05-31 #4362 @lambdaloop
Here is my review of the paper. It doesn't really fit as an issue in the Pose2Sim repository. Let me know if this should be placed somewhere else.

## Summary
> Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The summary is currently overly technical for a non-specialist audience. In particular, OpenPose and OpenSim are very specific tools that most people would not have heard of. It would be better to describe them, such as:
"from an OpenPose input, namely the 2D positions of joints from videos, to an OpenSim result, namely the position of a 3D skeletal model"
(it should be rephrased by the authors to match their view of this software)

I think the reference "User/Config.toml" file is too much detail for the Summmary.

## A statement of need
> Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

The paper briefly describes why markerless tracking is useful, but this is not actually the need that the package meets! The purpose of the software is in the title, to translate OpenPose outputs OpenSim outputs. This is indeed a useful contribution, but the reason is buried in the current statement. As it reads currently, it seems that the best contribution is you could obtain 3D angles from the OpenSim output. Why is this useful? It would help to detail why 3D angles are useful above 3D positions. Furthermore, as I understand, OpenSim allows you to further estimate forces, constrain movements biomechanically, produce cool animations of skeletons, and perhaps even provide a framework for interpreting the motion in terms of muscle activities. Very little of this is clear from the statement of need, and the paper would really benefit from reframing the full statement to make the advantage of bringing kinematic data into OpenSim clearer.

## State of the field
> Do the authors describe how this software compares to other commonly-used packages?

There is really not much detail about how the other packages compare. There is a sentence about Theia3D and Anipose computing 3D angles, but no further description. As a user, why should I use Pose2Sim over these other packages? There should be a longer description comparing these packages, even at least a sentence about how Pose2Sim does something that those do not.

## Quality of writing
> Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper has several typos. Generally, the whole paper could benefit from another pass to make it more clear and concise. Examples of some of these issues:
- "fall-off" on line 23 should be "fall off".
- "...networks are has been growing" on line 27 should be "has been growing"
- "it is allegedly easy to learn" on line 37 -- remove "allegedly" for clearer sentence
- "the 3D point coordinates are finally constrained" on lines 37/38 --> remove "finally"
- "AniPose" should be "Anipose" (line 31)

## References
> Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

- The Pagnon et al 2022 paper seems to be missing a title?
- I think it would help to cite a paper or two about how someone would estimate 3D angles with a marker-based approach in the statement of need.
- The OpenCV paper should be cited
