### 2022-05-31 #4362 @lambdaloop
For functionality, I was able to reproduce the run the demo after fixing the issues pointed out in the issues above.

I currently only use Linux and as far as I can tell I need to compile OpenSim from source to use it, so I wasn't able to test this part of the package. I hope one of the other reviewers can address this in particular!

I'll conclude by saying that bringing markerless 3D tracking data into OpenSim is something I've thought about for a while but didn't know how to do, so I'm excited to see this open-source pipeline to achieve this!
