+++ README.md
@@ -1,0 +1,2 @@
+# pose2sim
+This Python repository offers a way to perform a markerless kinematic analysis from multiple calibrated views as a unified workflow from an OpenPose input to an OpenSim result.
