+++ Content/paper.md
@@ -1,0 +1,136 @@
+---
+title: 'Pose2Sim: An Open Source Python Package for multiview markerless kinematics'
+tags:
+  - python
+  - markerless kinematics
+  - motion capture
+  - sports performance analysis
+  - openpose
+  - opensim
+authors:
+  - name: David Pagnon^[corresponding author] 
+    orcid: 0000-0002-6891-8331
+    affiliation: 1, 2
+  - name: Mathieu Domalain
+    orcid: 0000-0002-4518-8479
+    affiliation: 2
+  - name: Lionel Reveret
+    orcid: 0000-0002-0810-5187
+    affiliation: 3
+affiliations:
+ - name: Laboratoire Jean Kuntzmann, Université Grenoble Alpes, 700 avenue Centrale, 38400 Saint Martin d’Hères, France
+   index: 1
+ - name: Institut Pprime, 2 Bd des Frères Lumière, 86360 Chasseneuil-du-Poitou, France
+   index: 2
+ - name: Inria Grenoble Rhône-Alpes, 38330 Montbonnot-Saint-Martin, France
+   index: 3
+date: January 24 2021
+bibliography: paper.bib
+---
+
+# Summary
+
+`Pose2Sim` offers a way to perform a markerless kinematic analysis of human movement from multiple 
+calibrated views under Python, from an Openpose [@Cao_2019] input to an OpenSim [@Delp_2007] result.
+
+The repository presents a framework to perform the following tasks:\
+- Calibrating cameras, \
+- Tracking the person viewed by the most cameras, in case of several detections by OpenPose,\
+- Triangulating 2D coordinates and storing them in a .trc file,\
+- Filtering the 3D coordinates,\
+- Scaling and running inverse kinematics via OpenSim.
+
+Each task is easily customizable through the 'User/Config.toml' file, and requires only moderate Python skills. Pose2Sim is accessible at [https://gitlab.inria.fr/perfanalytics/pose2sim](https://gitlab.inria.fr/perfanalytics/pose2sim). 
+
+# Statement of need
+
+For the last few decades, marker-based kinematics has been considered as the best choice for the analysis of human movement, when regarding the trade-off between ease-of-use and accuracy. However, it cannot be considered as a gold-standard. Markers can be misplaced, move, or even fall-off, which introduces errors. Moreover, the system is hard to set outside or in "ecological" conditions, and it requires placing markers on the body, which can hinder natural movement. 
+
+The emergence of markerless kinematics opens up new possibilities. Indeed, the interest in deep learning pose estimation neural networks are has been growing fast since 2015 [@Zheng_2022], which makes it now possible to collect accurate and reliable kinematic data without the use of physical markers. Little work has been done towards obtaining 3D angles from multiple views [@Zheng_2022], aside from Theia3D [@Kanko_2021] as a commercial software, and AniPose [@Karashchuk_2021] in the animal motion analysis field.
+
[…]
