+++ Content/paper.bib
@@ -1,0 +1,119 @@
+@article{Cao_2019,
+  title={OpenPose: realtime multi-person 2D pose estimation using Part Affinity Fields},
+  author={Cao, Zhe and Hidalgo, Gines and Simon, Tomas and Wei, Shih-En and Sheikh, Yaser},
+  journal={IEEE transactions on pattern analysis and machine intelligence},
+  volume={43},
+  number={1},
+  pages={172--186},
+  year={2019},
+  URL = {https://arxiv.org/abs/1611.08050},
+  DOI = {10.1109/TPAMI.2019.2929257},
+  publisher={IEEE}
+}
+
+@article{Delp_2007,
+  title={OpenSim: open-source software to create and analyze dynamic simulations of movement},
+  author={Delp, Scott L and Anderson, Frank C and Arnold, Allison S and Loan, Peter and Habib, Ayman and John, Chand T and Guendelman, Eran and Thelen, Darryl G},
+  journal={IEEE transactions on biomedical engineering},
+  volume={54},
+  number={11},
+  pages={1940--1950},
+  year={2007},
+  URL = {https://ieeexplore.ieee.org/abstract/document/4352056},
+  DOI = {10.1109/TBME.2007.901024},
+  publisher={IEEE}
+}
+
+@inproceedings{Fang_2017,
+  title={{RMPE}: Regional Multi-person Pose Estimation},
+  author={Fang, Hao-Shu and Xie, Shuqin and Tai, Yu-Wing and Lu, Cewu},
+  booktitle={ICCV},
+  year={2017},
+  URL = {https://ieeexplore.ieee.org/document/8237518},
+  DOI = {10.1109/ICCV.2017.256}
+}
+
+@article{Kanko_2021,
+  title={Concurrent assessment of gait kinematics using marker-based and markerless motion capture},
+  author={Kanko, Robert M and Laende, Elise K and Davis, Elysia M and Selbie, W Scott and Deluzio, Kevin J},
+  journal={Journal of biomechanics},
+  volume={127},
+  pages={110665},
+  year={2021},
+  URL = {https://doi.org/10.1016/j.jbiomech.2021.110665},
+  DOI = {10.1016/j.jbiomech.2021.110665},
+  publisher={Elsevier}
+}
+
+@article{Karashchuk_2021,
+  title={Anipose: a toolkit for robust markerless 3D pose estimation},
+  author={Karashchuk, Pierre and Rupp, Katie L and Dickinson, Evyn S and Walling-Bell, Sarah and Sanders, Elischa and Azim, Eiman and Brunton, Bingni W and Tuthill, John C},
+  journal={Cell reports},
+  volume={36},
+  number={13},
+  pages={109730},
+  year={2021},
+  URL = {https://doi.org/10.1016/j.celrep.2021.109730},
+  DOI = {10.1016/j.celrep.2021.109730},
+  publisher={Elsevier}
+}
+
+@article{Mathis_2018,
+  title={DeepLabCut: markerless pose estimation of user-defined body parts with deep learning},
+  author={Mathis, Alexander and Mamidanna, Pranav and Cury, Kevin M and Abe, Taiga and Murthy, Venkatesh N and Mathis, Mackenzie Weygandt and Bethge, Matthias},
+  journal={Nature neuroscience},
+  volume={21},
+  number={9},
+  pages={1281--1289},
+  year={2018},
+  URL = {https://www.nature.com/articles/s41593-018-0209-y},
+  DOI = {10.1038/s41593-018-0209-y},
+  publisher={Nature Publishing Group}
+}
+
+@article{Pagnon_2021,
[…]
