+++ Pose2Sim/Utilities/calib_qca_to_toml.py
@@ -1,0 +1,210 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+'''
+    ##################################################
+    ## QCA CALIBRATION TO TOML CALIBRATION          ##
+    ##################################################
+    
+    Convert a Qualisys .qca.txt calibration file 
+    to an OpenCV .toml calibration file
+
+    Usage: 
+        from Pose2Sim.Utilities import calib_qca_to_toml; calib_qca_to_toml.calib_qca_to_toml_func(r'<input_qca_file>')
+        OR python -m calib_qca_to_toml -i "<input_qca_file>"
+        OR python -m calib_qca_to_toml -i "<input_qca_file>" --binning_factor 2 -o "<output_toml_file>"
+'''
+
+
+## INIT
+import os
+import argparse
+import numpy as np
+from lxml import etree
+import cv2
+
+
+## AUTHORSHIP INFORMATION
+__author__ = "David Pagnon"
+__copyright__ = "Copyright 2021, Pose2Sim"
+__credits__ = ["David Pagnon"]
+__license__ = "BSD 3-Clause License"
+__version__ = "0.1"
+__maintainer__ = "David Pagnon"
+__email__ = "contact@david-pagnon.com"
+__status__ = "Development"
+
+
+## FUNCTIONS
+def read_qca(qca_path, binning_factor):
+    '''
+    Read a Qualisys .qca.txt calibration file
+    Returns 5 lists of size N (N=number of cameras):
+    - C (camera name),
+    - S (image size),
+    - D (distorsion), 
+    - K (intrinsic parameters),
+    - R (extrinsic rotation),
+    - T (extrinsic translation)
+    '''
+
+    root = etree.parse(qca_path).getroot()
+    C, S, D, K, R, T = [], [], [], [], [], []
+
+    # Cameraa name
+    for tag in root.findall('cameras/camera'):
+        C += [tag.attrib.get('serial')]
+
+    # Image size
+    for tag in root.findall('cameras/camera/fov_marker'):
+        w = float(tag.attrib.get('right'))/binning_factor
+        h = float(tag.attrib.get('bottom'))/binning_factor
+        S += [[w, h]]
+    
+    # Intrinsic parameters: distorsion and intrinsic matrix
+    for tag in root.findall('cameras/camera/intrinsic'):
+        k1 = float(tag.get('radialDistortion1'))/64/binning_factor
+        k2 = float(tag.get('radialDistortion2'))/64/binning_factor
+        p1 = float(tag.get('tangentalDistortion1'))/64/binning_factor
+        p2 = float(tag.get('tangentalDistortion2'))/64/binning_factor
+        D+= [np.array([k1, k2, p1, p2])]
+        
+        fu = float(tag.get('focalLengthU'))/64/binning_factor
+        fv = float(tag.get('focalLengthV'))/64/binning_factor
+        cu = float(tag.get('centerPointU'))/64/binning_factor
+        cv = float(tag.get('centerPointV'))/64/binning_factor
+        K += [np.array([fu, 0., cu, 0., fv, cv, 0., 0., 1.]).reshape(3,3)]
+
+    # Extrinsic parameters: rotation matrix and translation vector
+    for tag in root.findall('cameras/camera/transform'):
+        tx = float(tag.get('x'))
+        ty = float(tag.get('y'))
+        tz = float(tag.get('z'))
+        r11 = float(tag.get('r11'))
[…]

[…]
