### 2022-06-09 #4362 @davidpagnon
Dear reviewers,
Thank you for your insightful reviews that we are going to take into account soon.
We have transferred the gitlab to this address: https://github.com/perfanalytics/pose2sim in order to make issues and contributing more coherent.
