+++ README.md
@@ -1,3 +1,6 @@
+<!-- continuous integration --> <!-- pypi version --><!-- downloads -->
+<!-- licence --> <!-- issues -->
+<!-- joss -->
 [![PyPI version](https://badge.fury.io/py/Pose2Sim.svg)](https://badge.fury.io/py/Pose2Sim)
 [![Downloads](https://pepy.tech/badge/pose2sim)](https://pepy.tech/project/pose2sim)
 [![License](https://img.shields.io/badge/License-BSD_3--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
