+++ setup.cfg
@@ -25,7 +25,7 @@ project_urls =
     
 [options]
 zip_safe = False
-python_requires = >=3.6, <3.9
+python_requires = >=3.6
 install_requires = 
     toml
     numpy
