+++ .github/workflows/python-package.yml
@@ -20,7 +20,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-11, windows-latest]
-        python-version: 3.8 #[3.6, 3.7, 3.8, 3.9, "3.10"]
+        python-version: [3.8]  #[3.6, 3.7, 3.8, 3.9, "3.10"]
         include:
           #- os: ubuntu-latest
           #  path: ~/.cache/pip
