+++ .github/workflows/python-package.yml
@@ -51,7 +51,5 @@ jobs:
           flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Test with pytest
         run: |
-          pip install pytest
-          pip install git+https://github.com/${{ github.repository }}.git@${{ github.sha }}
           cd Pose2Sim/Demo
           pytest Demo.py
