### 2022-06-12 415cdbc
install pyqt5 before matplotlib
