+++ setup.cfg
@@ -31,13 +31,13 @@ install_requires =
     numpy
     opencv-python
     lxml
+    PyQt5
     matplotlib
     tqdm
     anytree
     pandas
     scipy
     statsmodels
-    PyQt5
     c3d
 packages = find:
 
