+++ .github/workflows/python-package.yml
@@ -19,7 +19,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-11, windows-latest]
+        os: [windows-latest] #[ubuntu-latest, macos-11, windows-latest]
         python-version: [3.8]  #[3.6, 3.7, 3.8, 3.9, "3.10"]
         include:
           #- os: ubuntu-latest
