### 2022-06-13 be867ce
Update setup.cfg
