+++ setup.cfg
@@ -31,8 +31,8 @@ install_requires =
     numpy
     opencv-python
     lxml
-    PyQt5
     matplotlib
+    PyQt5
     tqdm
     anytree
     pandas
