+++ .github/workflows/python-package.yml
@@ -42,7 +42,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install flake8 pytest
-          python setup.py install
+          python -m pip install git+https://github.com/${{ github.repository }}.git@${{ github.sha }}
       - name: Lint with flake8
         run: |
           # stop the build if there are Python syntax errors or undefined names
