+++ Pose2Sim/Demo/Demo.py
@@ -1,5 +1,0 @@
-from Pose2Sim import Pose2Sim
-Pose2Sim.calibrateCams()
-Pose2Sim.track2D()
-Pose2Sim.triangulate3D()
-Pose2Sim.filter3D()
