+++ Pose2Sim/Demo/test.py
@@ -1,6 +1,8 @@
 import os
-from Pose2Sim import Pose2Sim
-Pose2Sim.calibrateCams(os.path.join('User', 'test.toml'))
-Pose2Sim.track2D(os.path.join('User', 'test.toml'))
-Pose2Sim.triangulate3D(os.path.join('User', 'test.toml'))
-Pose2Sim.filter3D(os.path.join('User', 'test.toml'))
+
+def workflow_test():
+  from Pose2Sim import Pose2Sim
+  Pose2Sim.calibrateCams(os.path.join('User', 'test.toml'))
+  Pose2Sim.track2D(os.path.join('User', 'test.toml'))
+  Pose2Sim.triangulate3D(os.path.join('User', 'test.toml'))
+  Pose2Sim.filter3D(os.path.join('User', 'test.toml'))
