+++ .github/workflows/python-package.yml
@@ -52,4 +52,4 @@ jobs:
       - name: Test with pytest
         run: |
           cd Pose2Sim/Demo
-          python -m test
+          pytest
