+++ .github/workflows/python-package.yml
@@ -51,5 +51,4 @@ jobs:
           flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Test with pytest
         run: |
-          cd Pose2Sim/Demo
           pytest
