+++ .github/workflows/python-package.yml
@@ -51,4 +51,5 @@ jobs:
           flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Test with pytest
         run: |
-          pytest
+          cd Pose2Sim/Demo
+          pytest -v test*
