+++ Pose2Sim/Demo/test.py
@@ -1,6 +1,6 @@
 import os
 
-def workflow_test():
+def test_workflow():
   from Pose2Sim import Pose2Sim
   Pose2Sim.calibrateCams(os.path.join('User', 'test.toml'))
   Pose2Sim.track2D(os.path.join('User', 'test.toml'))
