### 2022-06-13 a70bf12
test also on ubuntu and macOS
