+++ .github/workflows/python-package.yml
@@ -19,13 +19,13 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [windows-latest] #[ubuntu-latest, macos-11, windows-latest]
+        os: [ubuntu-latest, macos-11, windows-latest]
         python-version: [3.8]  #[3.6, 3.7, 3.8, 3.9, "3.10"]
         include:
-          #- os: ubuntu-latest
-          #  path: ~/.cache/pip
-          #- os: macos-11
-          #  path: ~/Library/Caches/pip
+          - os: ubuntu-latest
+            path: ~/.cache/pip
+          - os: macos-11
+            path: ~/Library/Caches/pip
           - os: windows-latest
             path: ~\AppData\Local\pip\Cache
 
