### 2022-06-13 88e521c
Add python 3.6 to 3.10
