### 2022-06-13 9ad153f
Test also on python 3.7 to 3.10
