### 2022-06-13 0271b4f
Update and rename python-package.yml to continuous-integration.yml
