### 2022-06-20 perfanalytics/pose2sim#9 @jonmatthis
JOSS Review Checklist - Comments and requests
I'm just going to leave a list of comments here rather than making a new issue for each thing, which seems like it blow up the number of issues and be hard to track and manage.

I'll use this issue as a rolling place to add my comments as I go. I'll make it clear when I'm "Done" with my first pass through (it might be a few days from this initial post)

Link to my review checklist comment in the official review  thread - https://github.com/openjournals/joss-reviews/issues/4362#issuecomment-1115354966
