### 2022-06-20 perfanalytics/pose2sim#9 @jonmatthis
# Repository: Is the source code for this software available at the https://gitlab.inria.fr/perfanalytics/pose2sim?

Obviously the repo was moved to this Github location :P
