### 2022-06-20 perfanalytics/pose2sim#9 @jonmatthis
# Installation

- Installing OpenPose is notoriously challenging, especially for new users.

Do they only need to use the "Windows Portable Demo?"

If so, you should specify that specifically. The link you sent will prompt many newbies to try to go through the full installation process for OpenPose (speaking from experience when I first discovered OpenPose in 2017 😅 )

If you DO need a full installation of OpenPose to run the software, you should specify that users CANNOT just use the Windows Portable Demo

- You should specify which versions of Open Sim have been tested (i.e. `2. Install OpenSim 4.x (tested with v4.3, download-link)
