### 2022-06-20 perfanalytics/pose2sim#9 @jonmatthis
# Installation cont

- You are missing the step where you tell people how to *activate* the conda environment with `conda activate Pose2Sim`

people will miss that step, install things to to their (base) install and then have a devil of a time figuring out why.

I mean, they'll do that regardless, but at least this way it'll be their fault :)
