### 2022-06-20 #4362 @jonmatthis
> * **Repository:** Is the source code for this software available at the https://gitlab.inria.fr/perfanalytics/pose2sim?

I'm including my review in the form of a comment thread in this Issue on the main repo - https://github.com/perfanalytics/pose2sim/issues/9
