### 2022-06-20 #4362 @jonmatthis
For the record (regarding the JOSS review process) -

 I wish I was able to open a comment thread from my own Checklist comment and include my comments there.

The method of "I'm Going to write everything up in a great big text blob and plop it in one big comment" feels inconvenient for all involved.

Maybe in the future, the check list could also auto-generate a linked Issue in the target repository?
