### 2022-06-20 perfanalytics/pose2sim#9 @jonmatthis
Minor -

When possible, Python based projects should follow the PEP8 style guide - https://peps.python.org/pep-0008/

In particular,  your `pip` install references `Pose2Sim` which is not using the recommended capitalization scheme. Technically you should name your pypi project `pose2sim`

(But feel free not to bother, just pointing out that you're not PEP8 compliant)
