### 2022-06-20 #4362 @jonmatthis
If your demonstration produces labeled images, it would be best to make those visible to the user (or at least  link to where they are located)
