### 2022-06-20 perfanalytics/pose2sim#9 @jonmatthis
# Installation

Ok, I've finished the Installation

It's mostly good, but for those relatively minor comments above.

You could help new users a lot by sign-posting that the "Use your own data" instructions are ALSO part of the "demonstation"

As listed, a person will start from the top, work through the demo, and then think that the rest of the ReadMe is for people who have recorded their own data.

You could provide the users with expected output from the "demonstration" commands and then explain that deeper explanation and instructions for loading the Sample data into OpenSim are presented below
