### 2022-06-21 #4362 @jonmatthis
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04362/joss.04362/10.21105.joss.04362.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.04362/joss.04362/10.21105.joss.04362.pdf) 📄 👈

These links are broken, where can I find this paper?
