### 2022-06-21 perfanalytics/pose2sim#9 @jonmatthis
final comments back on main review thread -

https://github.com/openjournals/joss-reviews/issues/4362#issuecomment-1161853931
