### 2022-07-04 c0c4952
Pose2Sim to pose2sim and changed issue url
remove capitals for PEP8 compliance
url changed from giltab to new github repo
