+++ setup.cfg
@@ -1,12 +1,12 @@
 [metadata]
-name = Pose2Sim
-version = 0.1.8
+name = pose2sim
+version = 0.1.9
 author = David Pagnon
 author_email = contact@david-pagnon.com
 description = Perform a markerless kinematic analysis from multiple calibrated views as a unified workflow from an OpenPose input to an OpenSim result.
 long_description = file: README.md, LICENSE
 long_description_content_type = text/markdown
-url = https://gitlab.inria.fr/perfanalytics/pose2sim
+url = https://github.com/perfanalytics/pose2sim
 keywords = markerless, kinematics, OpenPose, OpenSim, 3D human pose, biomechanics
 license = BSD 3-Clause License
 license_files = LICENSE
@@ -21,7 +21,7 @@ classifiers =
     Topic :: Scientific/Engineering :: Image Processing
     Topic :: Scientific/Engineering :: Medical Science Apps.
 project_urls = 
-    Bug Tracker = https://github.com/davidpagnon/pose2sim/issues
+    Bug Tracker = https://github.com/perfanalytics/pose2sim/issues
     
 [options]
 zip_safe = False
