### 2022-07-04 perfanalytics/pose2sim#9 @davidpagnon
Dear Jonathan,
Thank you for your review!

> # Installation
> * Installing OpenPose is notoriously challenging, especially for new users.
> Do they only need to use the "Windows Portable Demo?"
> If so, you should specify that specifically. The link you sent will prompt many newbies to try to go through the full installation process for OpenPose (speaking from experience when I first discovered OpenPose in 2017 😅 )
> If you DO need a full installation of OpenPose to run the software, you should specify that users CANNOT just use the Windows Portable Demo

> * You should specify which versions of Open Sim have been tested (i.e. `2. Install OpenSim 4.x (tested with v4.3, download-link)

I now specified these:
1. **Install OpenPose** (instructions [there](https://github.com/CMU-Perceptual-Computing-Lab/openpose/blob/master/doc/installation/0_index.md)). Portable demo is enough.
2. **Install OpenSim 4.x** ([there](https://simtk.org/frs/index.php?group_id=91)). Tested up to v4.4-beta.

> * You are missing the step where you tell people how to _activate_ the conda environment with `conda activate Pose2Sim`
> people will miss that step, install things to to their (base) install and then have a devil of a time figuring out why.

I don't think I forgot this step, unless I am missing something?
<pre><i>conda create -n Pose2Sim python=3.8.8
conda activate Pose2Sim</i></pre>

> Your instructions currently make it look like you should install Pose2Sim from `pip` and THEN build from source. You should specify that it is one or the other

I see, I made it clear now that it's two different options.

> your `pip` install references `Pose2Sim` which is not using the recommended capitalization scheme. Technically you should name your pypi project `pose2sim`

You are right, and it is case insensitive anyway. Next PyPI release will be named pose2sim (see setup.cfg).

> the command `pip show Pose2Sim | grep Location` does not work on my computer because it doesn't recognize the `grep` command.
What kind of terminal are you using? I'm a little suprised, as it works with me on linux shell, powershell, anaconda prompt.
`pip show pose2sim` would work, however you would have to find the "Location"  by yourself.

I'll post this for now, before I make a mistake and lose all of my answers :sweat_smile:
