### 2022-07-04 perfanalytics/pose2sim#9 @davidpagnon
> `Open a python session in your Pose2Sim\Demo folder, and test the following code:`
> Statements like this are inscrutable to new-comers. Give them specific instructions on how to `cd` into that folder and start a python session by entering the command `python`

This has been made clearer for beginners: \
Open a terminal and check package location with `pip show pose2sim | grep Location`. \
Copy this path and go to the Demo folder with `cd <path>\pose2sim\Demo`. \
Open a python session with `python` command, and test the following code

-----

> You could help new users a lot by sign-posting that the "Use your own data" instructions are ALSO part of the "demonstation"
> As listed, a person will start from the top, work through the demo, and then think that the rest of the ReadMe is for people who have recorded their own data.
> You could provide the users with expected output from the "demonstration" commands and then explain that deeper explanation and instructions for loading the Sample data into OpenSim are presented below

I made it clearer in the caption of the Demonstration section: \
_**This demonstration provides an example experiment of a person balancing on a beam, filmed with 4 calibrated cameras processed with OpenPose. Deeper explanations and instructions are given on the section below.**_

 -----

> also -If your demonstration produces labeled images, it would be best to make those visible to the user (or at least link to where they are located

My demonstration doesn't provide labeled images, unfortunately my lab's policy doesn't want anyone to be recognizable on a public repository (I actually had the hardest time even being allowed to provide any json files at all.)

I'm going to work on a Maya and/or Blender add-on to get a visual feedback at all stages, but this is out of the scope of this current review.

Thank you for your comments and suggestions!
