+++ README.md
@@ -544,25 +544,22 @@ If you use this code or data, please cite [Pagnon et al., 2022](https://www.mdpi
 
 #### How to contribute
 
-I am currently the sole developer of Pose2Sim, but I would happily welcome any proposal for new features, code improvement, etc!\
-If you want to contribute, please follow [this guide](https://docs.github.com/en/get-started/quickstart/contributing-to-projects) on how to fork, modify and push code, and submit a pull request. \
-I would appreciate it if you provided as much useful information as possible about how you modified the code, and a rationale for why you're making this pull request.\
+I would happily welcome any proposal for new features, code improvement, and more!\
+If you want to contribute to Pose2Sim, please follow [this guide](https://docs.github.com/en/get-started/quickstart/contributing-to-projects) on how to fork, modify and push code, and submit a pull request. I would appreciate it if you provided as much useful information as possible about how you modified the code, and a rationale for why you're making this pull request.\
 Please also specify on which operating system and on which python version you have tested the code.
 
-<details>
-  <summary><b>To-do list, for general guidance purposes only</b> (CLICK TO SHOW)</summary>
-<li> Integrate as a Blender and / or Maya add-on. See <a href="https://github.com/davidpagnon/Maya-Mocap">Maya-Mocap</a> and <a href="https://github.com/JonathanCamargo/BlendOsim">BlendOSim</a>
-<li> Multiple persons kinematics (triangulating multiple persons, and sorting them in time)</li>
-<li> Use aniposelib for better calibration </li>
+*Here is a to-do list, for general guidance purposes only:*
+> <li> Integrate as a Blender and / or Maya add-on. See <a href="https://github.com/davidpagnon/Maya-Mocap">Maya-Mocap</a> and <a href="https://github.com/JonathanCamargo/BlendOsim">BlendOSim</a>
+> <li> Multiple persons kinematics (triangulating multiple persons, and sorting them in time)</li>
+> <li> Use aniposelib for better calibration </li></br>
+> 
+> <li> Conda package and Docker image</li>
+> <li> Solve limb swapping</li>
+> <li> Implement normalized DLT and RANSAC triangulation, as well as a triangulation refinement step (cf DOI:10.1109/TMM.2022.3171102)</li>
+> <li> Implement optimal fixed-interval Kalman smoothing</li>
+> <li> Utilities: convert Vicon xcp calibration file to toml</li>
+> <li> Run from command line via click or typer</li>
+> <li> Catch errors</li>
+> <li> Make GUI</li>
 
 
-<li> Conda package and Docker image</li>
-<li> Solve limb swapping</li>
-<li> Implement normalized DLT and RANSAC triangulation, as well as a triangulation refinement step (cf DOI:10.1109/TMM.2022.3171102)</li>
-<li> Implement optimal fixed-interval Kalman smoothing</li>
-<li> Utilities: convert Vicon xcp calibration file to toml</li>
-<li> Run from command line via click or typer</li>
-<li> Catch errors</li>
-
-
-</details>
