+++ README.md
@@ -7,14 +7,16 @@
 [![status](https://joss.theoj.org/papers/a31cb207a180f7ac9838d049e3a0de26/status.svg)](https://joss.theoj.org/papers/a31cb207a180f7ac9838d049e3a0de26)
 
 # Pose2Sim
-This Python repository offers a way to perform a markerless kinematic analysis from multiple calibrated views as a unified workflow from an OpenPose input to an OpenSim result.
+`Pose2Sim` provides a workflow for 3D markerless kinematics, as an alternative to the more usual marker-based motion capture methods.\
+Pose2Sim stands for "OpenPose to OpenSim", as it uses OpenPose inputs (2D keypoints coordinates obtained from multiple videos) and leads to an OpenSim result (full-body 3D joint angles). 
 
 <img src="Content/Overview.png" width="760">
 
 ## Contents
 1. [Installation and Demonstration](#installation-and-demonstration)
    1. [Installation](#installation)
-   2. [Demonstration](#demonstration)
+   2. [Demonstration Part-1: Build 3D TRC file on Python](#demonstration-part-1-build-3d-trc-file-on-python)
+   3. [Demonstration Part-2: Obtain 3D joint angles with OpenSim](#demonstration-part-2-obtain-3d-joint-angles-with-opensim)
 2. [Use on your own data](#use-on-your-own-data)
    1. [Prepare for running on your own data](#prepare-for-running-on-your-own-data)
    2. [2D pose estimation](#2d-pose-estimation)
@@ -53,12 +55,12 @@ If you don't use Anaconda, type `python -V` in terminal to make sure python>=3.6
        pip install .
        ```
           
-### Demonstration   
-> _**This demonstration provides an example experiment of a person balancing on a beam, filmed with 4 calibrated cameras processed with OpenPose. Deeper explanations and instructions are given on the section below.**_ 
+### Demonstration Part-1: Build 3D TRC file on Python  
+> _**This demonstration provides an example experiment of a person balancing on a beam, filmed with 4 calibrated cameras processed with OpenPose.**_ 
 
 Open a terminal and check package location with `pip show pose2sim | grep Location`. \
-Copy this <path> and go to the Demo folder with `cd <path>\pose2sim\Demo`. \
-Open a python session with `python` command, and test the following code:
+Copy this path and go to the Demo folder with `cd <path>\pose2sim\Demo`. \
+Type `python`, and test the following code:
 ```
 from Pose2Sim import Pose2Sim
 Pose2Sim.calibrateCams()
@@ -68,11 +70,27 @@ Pose2Sim.filter3D()
 ```
 You should obtain a plot of all the 3D coordinates trajectories. You can check the logs in`Demo\Users\logs.txt`.\
 Results are stored as .trc files in the `Demo/pose-3d` directory. \
-Section [Opensim kinematics](#opensim-kinematics) describes how to obtain physically accurate 3D joint angles.
 
-*N.B.:* Default parameters have been provided in `Demo\Users\Config.toml` but can be edited.\
+*N.B.:* Default parameters have been provided in `Demo\Users\Config.toml` but can be edited.
+<br/>
+
+### Demonstration Part-2: Obtain 3D joint angles with OpenSim  
+> _**In the same vein as you would do with marker-based kinematics, start with scaling your model, and then perform inverse kinematics.**_ 
+
+#### Scaling
+1. Open OpenSim.
+2. Open the provided `Model_Pose2Sim_Body25b.osim` model from `pose2sim/Demo/opensim`. *(File -> Open Model)*
+3. Load the provided `Scaling_Setup_Pose2Sim_Body25b.xml` scaling file from `pose2sim/Demo/opensim`. *(Tools -> Scale model -> Load)*
+4. Run. You should see your skeletal model take the static pose.
+
[…]
