+++ Content/paper.md
@@ -30,27 +30,29 @@ bibliography: paper.bib
 
 # Summary
 
-`Pose2Sim` offers a way to perform a markerless kinematic analysis of human movement from multiple 
-calibrated views under Python, from an Openpose [@Cao_2019] input to an OpenSim [@Delp_2007] result.
+`Pose2Sim` provides a workflow for 3D markerless kinematics, as an alternative to the more usual marker-based motion capture methods.\
+Pose2Sim stands for "OpenPose to OpenSim", as it uses OpenPose inputs (2D coordinates obtained from multiple videos) and leads to an OpenSim result (full-body 3D joint angles). 
 
-The repository presents a framework to perform the following tasks:\
+The repository presents a framework for:\
+- Detecting 2D joint coordinates from videos, e.g. via OpenPose [@Cao_2019],\ 
 - Calibrating cameras, \
-- Tracking the person viewed by the most cameras, in case of several detections by OpenPose,\
-- Triangulating 2D coordinates and storing them in a .trc file,\
-- Filtering the 3D coordinates,\
-- Scaling and running inverse kinematics via OpenSim.
+- Tracking the main person on the scene,\
+- Triangulating 2D joint coordinates and storing them as 3D positions in a .trc file,\
+- Filtering these calculated 3D positions,\
+- Scaling and running inverse kinematics via OpenSim [@Delp_2007; @Seth_2018], in order to obtain full-body 3D joint angles.
 
-Each task is easily customizable through the 'User/Config.toml' file, and requires only moderate Python skills. Pose2Sim is accessible at [https://gitlab.inria.fr/perfanalytics/pose2sim](https://gitlab.inria.fr/perfanalytics/pose2sim). 
+Each task is easily customizable, and requires only moderate Python skills. Pose2Sim is accessible at [https://github.com/perfanalytics/pose2sim](https://github.com/perfanalytics/pose2sim). 
 
 # Statement of need
 
-For the last few decades, marker-based kinematics has been considered as the best choice for the analysis of human movement, when regarding the trade-off between ease-of-use and accuracy. However, it cannot be considered as a gold-standard. Markers can be misplaced, move, or even fall-off, which introduces errors. Moreover, the system is hard to set outside or in "ecological" conditions, and it requires placing markers on the body, which can hinder natural movement. 
+For the last few decades, marker-based kinematics has been considered as the best choice for the analysis of human movement, when regarding the trade-off between ease-of-use and accuracy. However, the system is hard to set outdoors or in "ecological" conditions, and it requires placing markers on the body, which can hinder natural movement. 
 
-The emergence of markerless kinematics opens up new possibilities. Indeed, the interest in deep learning pose estimation neural networks are has been growing fast since 2015 [@Zheng_2022], which makes it now possible to collect accurate and reliable kinematic data without the use of physical markers. Little work has been done towards obtaining 3D angles from multiple views [@Zheng_2022], aside from Theia3D [@Kanko_2021] as a commercial software, and AniPose [@Karashchuk_2021] in the animal motion analysis field.
[…]
