+++ Content/paper.bib
@@ -1,3 +1,14 @@
+@article{Bradski_2000,
+    author = {Bradski, G.},
+    citeulike-article-id = {2236121},
+    journal = {Dr. Dobb's Journal of Software Tools},
+    keywords = {bibtex-import},
+    posted-at = {2008-01-15 19:21:54},
+    priority = {4},
+    title = {{The OpenCV Library}},
+    year = {2000}
+}
+
 @article{Cao_2019,
   title={OpenPose: realtime multi-person 2D pose estimation using Part Affinity Fields},
   author={Cao, Zhe and Hidalgo, Gines and Simon, Tomas and Wei, Shih-En and Sheikh, Yaser},
@@ -84,15 +95,30 @@
 }
 
 @article{Pagnon_2022,
-  AUTHOR = {Pagnon, David and Domalain, Mathieu and Reveret, Lionel},
-  TITEL = {Pose2Sim: An End-to-End Workflow for 3D Markerless Sports Kinematics—Part 2: Accuracy},
-  JOURNAL = {Sensors},
+  title = {Pose2Sim: An End-to-End Workflow for 3D Markerless Sports Kinematics—Part 2: Accuracy},
+  author = {Pagnon, David and Domalain, Mathieu and Reveret, Lionel},
+  journal = {Sensors},
   volume={22},
   number={7},
   year={2022},
   URL = {https://www.mdpi.com/1424-8220/22/7/2712},
   DOI = {10.3390/s22072712},
   publisher={Multidisciplinary Digital Publishing Institute}
+}
+
+@article{Seth_2018,
+    doi = {10.1371/journal.pcbi.1006223},
+    author = {Seth, Ajay AND Hicks, Jennifer L. AND Uchida, Thomas K. AND Habib, Ayman AND Dembia, Christopher L. AND Dunne, James J. AND Ong, Carmichael F. AND DeMers, Matthew S. AND Rajagopal, Apoorva AND Millard, Matthew AND Hamner, Samuel R. AND Arnold, Edith M. AND Yong, Jennifer R. AND Lakshmikanth, Shrinidhi K. AND Sherman, Michael A. AND Ku, Joy P. AND Delp, Scott L.},
+    journal = {PLOS Computational Biology},
+    publisher = {Public Library of Science},
+    title = {OpenSim: Simulating musculoskeletal dynamics and neuromuscular control to study human and animal movement},
+    year = {2018},
+    month = {07},
+    volume = {14},
+    url = {https://doi.org/10.1371/journal.pcbi.1006223},
+    pages = {1-20},
[…]
