+++ Content/paper.md
@@ -34,12 +34,12 @@ bibliography: paper.bib
 Pose2Sim stands for "OpenPose to OpenSim", as it uses OpenPose inputs (2D coordinates obtained from multiple videos) and leads to an OpenSim result (full-body 3D joint angles). 
 
 The repository presents a framework for:\
-- Detecting 2D joint coordinates from videos, e.g. via OpenPose [@Cao_2019],\ 
-- Calibrating cameras, \
-- Tracking the main person on the scene,\
-- Triangulating 2D joint coordinates and storing them as 3D positions in a .trc file,\
-- Filtering these calculated 3D positions,\
-- Scaling and running inverse kinematics via OpenSim [@Delp_2007; @Seth_2018], in order to obtain full-body 3D joint angles.
+• Detecting 2D joint coordinates from videos, e.g. via OpenPose [@Cao_2019],\ 
+• Calibrating cameras, \
+• Tracking the main person on the scene,\
+• Triangulating 2D joint coordinates and storing them as 3D positions in a .trc file,\
+• Filtering these calculated 3D positions,\
+• Scaling and running inverse kinematics via OpenSim [@Delp_2007; @Seth_2018], in order to obtain full-body 3D joint angles.
 
 Each task is easily customizable, and requires only moderate Python skills. Pose2Sim is accessible at [https://github.com/perfanalytics/pose2sim](https://github.com/perfanalytics/pose2sim). 
 
@@ -56,9 +56,11 @@ So far, little work has been done towards obtaining 3D angles from multiple view
 
 # Features
 ## Pose2Sim workflow
+![Pose2Sim full pipeline: (1) OpenPose 2D joint detection; (2i) Camera calibration; (2ii–iv) Tracking the person of interest, Triangulating his coordinates, and Filtering them; (3) Constraining the 3D coordinates to a physically consistent OpenSim skeletal model.\label{fig:Pose2Sim pipeline}](Pipeline.png)
+
 `Pose2Sim` connects two of the most widely recognized (and open source) pieces of software of their respective fields:\
-- OpenPose [@Cao_2019], a 2D human pose estimation neural network\
-- OpenSim [@Delp_2007], a 3D biomechanics analysis software
+• OpenPose [@Cao_2019], a 2D human pose estimation neural network\
+• OpenSim [@Delp_2007], a 3D biomechanics analysis software
 
 The workflow is organized as follows:\
 1. Preliminary OpenPose [@Cao_2019] 2D keypoint detection.\
@@ -71,64 +73,62 @@ The workflow is organized as follows:\
 
 OpenPose, OpenSim, as well as the whole `Pose2Sim` workflow run from any video cameras, on any computer, equipped with any operating system. 
 
-![Pose2Sim pipeline.\label{fig:Pose2Sim full pipeline: (1) OpenPose 2D joint detection; (2i) Camera calibration; (2ii–iv) Tracking the person of interest, Triangulating his coordinates, and Filtering them; (3) Constraining the 3D coordinates to a physically consistent OpenSim skeletal model.}](Pipeline.png)
-
 
 ## Pose2Sim core
 Each step of the Pose2Sim core is easily customizable through the 'User/Config.toml' file. Among other things, users can edit:\
-- The project hierarchy, the video framerate, the range of analyzed frames,\
-- The OpenPose model they wish to use. They can also use AlphaPose [@Fang_2017], or even create their own model (e.g. with DeepLabCut [@Mathis_2018]),\
-- Whether they are going to calibrate their cameras with a checkerboard, or to simply convert a calibration file provided by a Qualisys system,\
-- Which keypoint they want to track in order to automatically single out the person of interest,\
[…]
