### 2022-07-08 8440b34
Add LJK to Lionel's affiliation
