### 2022-07-08 b855ea1
Ignore changes in Readme.md and in Content folder
