### 2022-07-15 #4362 @davidpagnon
Dear reviewers,

Thank you for your input! I'm sorry about my late answer, I had a difficult deadline to meet. Anyway, here it is!\
I answered to the issues you have posted on the repository (feel free to follow up if you're not satisfied), now here are my answers to each of the reviewers in separate posts.

Also, I'm not sure you're going to be able to compile the paper with the @editorialbot generate pdf since we've changed directory, so please find attached the pdf.

[paper_post.pdf](https://github.com/openjournals/joss-reviews/files/9124697/paper_post.pdf)
