+++ .github/workflows/joss_pdf.yml
@@ -18,7 +18,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: Content/paper.md
+          paper-path: Content/paper_joss.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -26,4 +26,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: Content/paper.pdf
+          path: Content/paper_joss.pdf
