+++ .github/workflows/joss_pdf.yml
@@ -27,4 +27,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: Content/paper_joss.pdf
+          path: Content/paper.pdf
