### 2022-07-16 #4362 @davidpagnon
Answers to @jonmatthis.

The other I answered your other feedback directly on this issue: https://github.com/perfanalytics/pose2sim/issues/9

> Specifically regarding the commentary on the use of marker-based motion capture as a "gold standard" -
> Marker-based motion capture is the gold standard for generating accurate recordings of human movement. Your concerns are completely valid, but the remains that marker-based mocap is the most accurate method currently available for recording 3d movement.
> I think you should alter that section to recognize that fact, but maintain your concerns about marker placement and the potential for innacurate measurements that could arise as a result. Basically something to the effect of "Marker based mocap is the most accurate method currently available, but it's got all theses problems so we should be trying to find better alternative solutions"

I rephrased this part to:
"For the last few decades, marker-based kinematics has been considered as the best choice for the analysis of human movement, when regarding the trade-off between ease-of-use and accuracy. However, the system is hard to set outdoors or in "ecological" conditions, and it requires placing markers on the body, which can hinder natural movement.
The emergence of markerless kinematics opens up new possibilities."

-----

> Please put a bit of polish into the formatting before publication
> A lot of the formatting in the main document is a bit sloppy. A lot of things that are supposed to be 'bullet points' but are in fact just hyphens, and no spaces between normal text and those not-quit-bulleted lists (e.g. Line 64-76) I'd like to see a bit more polish in the formatting before final acceptance, if possible.

Bullet points now replace hyphens.

> Figure 1 needs a caption, ideally providing a brief summary of what is happening in each panel.

The caption is now more explanatory:
"Pose2Sim full pipeline: (1) OpenPose 2D joint detection; (2i) Camera calibration; (2ii–iv) Tracking the person of interest, Triangulating his coordinates, and Filtering them; (3) Constraining the 3D coordinates to a physically consistent OpenSim skeletal model."

> Also, maybe draw a border between the images? At the moment it is very difficult to parse what I'm seeing. e.g. in the upper right panel - is that a bunch of people in the same area? or multiple images super imposed on each other?
> I wouldn't mind if you broke Figure 1 up into multiple figures, if that makes it easier to clarify what is happening in each panel

I drew a clearer border, and with the caption I believe the figure is clear enough. Feel free to tell me if you still need to break it into several figures!
