+++ README.md
@@ -10,7 +10,7 @@
 `Pose2Sim` provides a workflow for 3D markerless kinematics, as an alternative to the more usual marker-based motion capture methods.\
 Pose2Sim stands for "OpenPose to OpenSim", as it uses OpenPose inputs (2D keypoints coordinates obtained from multiple videos) and leads to an OpenSim result (full-body 3D joint angles). 
 
-<img src="Content/Pose2Sim_workflow.png" width="760">
+<img src="Content/Pose2Sim_workflow.jpg" width="760">
 <img src="Content/Overview.png" width="760">
 
 
