### 2022-07-20 b011517
remove pipeline.png
