### 2022-07-28 5c2692f
natural sorting of json files
folders were correctly sorted, but not always json files. 
Example: if the files are named 1.json, 2.json, ..., 10.json, they were ordered as 1.json, 10.json, 2.json, ...
