+++ Pose2Sim/triangulate_3d.py
@@ -384,9 +384,11 @@ def triangulate_all(config):
     json_dirs_names = [k for k in pose_listdirs_names if json_folder_extension in k]
     try: 
         json_files_names = [fnmatch.filter(os.listdir(os.path.join(poseTracked_dir, js_dir)), '*.json') for js_dir in json_dirs_names]
+        json_files_names = [natural_sort(j) for j in json_files_names]
         json_tracked_files = [[os.path.join(poseTracked_dir, j_dir, j_file) for j_file in json_files_names[j]] for j, j_dir in enumerate(json_dirs_names)]
     except:
         json_files_names = [fnmatch.filter(os.listdir(os.path.join(pose_dir, js_dir)), '*.json') for js_dir in json_dirs_names]
+        json_files_names = [natural_sort(j) for j in json_files_names]
         json_tracked_files = [[os.path.join(pose_dir, j_dir, j_file) for j_file in json_files_names[j]] for j, j_dir in enumerate(json_dirs_names)]
     
     # Triangulation
