### 2022-07-29 fbc3161
natural sorting of json files
folders were correctly sorted, but not always json files. 
Example: if the files are named 1.json, 2.json, ..., 10.json, they were ordered as 1.json, 10.json, 2.json, ...
