### 2022-07-29 2a75ee5
Create combine_trc.py
