+++ Pose2Sim/Utilities/combine_trc.py
@@ -1,0 +1,168 @@
+#! /usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+'''
+    ##################################################
+    ## Combine two trc files                        ##
+    ##################################################
+    
+    Combine two trc files.
+    Example: you have run Pose2Sim with OpenPose AND with a DeepLabCut model 
+    (or any other marker-based or markerless pose estimation algorithm), 
+    and you want to assemble both detections before running OpenSim.
+    
+    Usage:
+    from Pose2Sim.Utilities import combine_trc; combine_trc.combine_trc_func(r'<first_path>', r'<second_path>', r'<output_path>')
+    OR python -m combine_trc -i "<first_path>" -j "<second_path>" -o "<output_path>"
+    OR python -m combine_trc -i "<first_path>" -j "<second_path>"
+'''
+
+
+## INIT
+import os
+import pandas as pd
+import numpy as np
+import argparse
+
+
+## AUTHORSHIP INFORMATION
+__author__ = "David Pagnon"
+__copyright__ = "Copyright 2022, Pose2Sim"
+__credits__ = ["David Pagnon"]
+__license__ = "BSD 3-Clause License"
+__version__ = "0.1"
+__maintainer__ = "David Pagnon"
+__email__ = "contact@david-pagnon.com"
+__status__ = "Development"
+
+
+## FUNCTIONS
+def df_from_trc(trc_path):
+    '''
+    Retrieve header and data from trc path.
+
+    INPUT:
+    trc_path: path to trc file
+
+    OUTPUT:
+    header: dictionary of header data
+    data: pandas dataframe of data
+    '''
+
+    # DataRate	CameraRate	NumFrames	NumMarkers	Units	OrigDataRate	OrigDataStartFrame	OrigNumFrames
+    df_header = pd.read_csv(trc_path, sep="\t", skiprows=1, header=None, nrows=2, encoding="ISO-8859-1")
+    header = dict(zip(df_header.iloc[0].tolist(), df_header.iloc[1].tolist()))
+    
+    # Label1_X  Label1_Y    Label1_Z    Label2_X    Label2_Y
+    df_lab = pd.read_csv(trc_path, sep="\t", skiprows=3, nrows=1)
+    labels = df_lab.columns.tolist()[2:-1:3]
+    labels_XYZ = np.array([[labels[i]+'_X', labels[i]+'_Y', labels[i]+'_Z'] for i in range(len(labels))], dtype='object').flatten()
+    labels_FTXYZ = np.concatenate((['Frame#','Time'], labels_XYZ))
+    
+    data = pd.read_csv(trc_path, sep="\t", skiprows=5, index_col=False, header=None, names=labels_FTXYZ)
+    
+    return header, data
+
+
+def combine_trc_headerdata (first_path, second_path):
+    '''
+    Combine headers and data from two different trc files.
+
+    INPUT:
+    first_path: path to first trc file
+    second_path: path to second trc file
+
+    OUTPUT:
+    Header: dictionary of combined headers
+    Data: dataframe of combined trc data
+    '''
+
+    first = df_from_trc(first_path)
+    second = df_from_trc(second_path)
+
+    frames_first = int(first[0].get('NumFrames'))
+    frames_second = int(second[0].get('NumFrames'))
+    NumFrames = min(frames_first, frames_second)
+    OrigNumFrames = NumFrames
+    NumMarkers = int(first[0].get('NumMarkers')) + int(second[0].get('NumMarkers'))
+
+    Header = first[0]
[…]
