### 2022-07-29 47e685c
Update and rename combine_trc.py to trc_combine.py
