### 2022-07-29 1830fdf
Add trc_combine in Utilities
