### 2022-08-11 #4362 @CVHammond
The changes made look great! The new GitHub repository looks professional and I like the Pose2Sim icon (whether it was made in jest or not, I won't know :sweat_smile:). Your responses to my concerns were more than satisfactory and I've updated my checklist.
