### 2022-08-11 #4362 @davidpagnon
Thank you all for your feedback, it was truly useful and pertinent!

@CVHammond, to be honest, I know close to nothing about Javascript, and I didn't even know about Jest. Do you refer to the testing, or to the icon? I drew the Icon with [Inkscape](https://inkscape.org/), and the tests with a [github action](https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python).
