### 2022-08-13 8aa0b99
version accepted for JOSS publication
