+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pose2sim
-version = 0.1.9
+version = 0.2.0
 author = David Pagnon
 author_email = contact@david-pagnon.com
 description = Perform a markerless kinematic analysis from multiple calibrated views as a unified workflow from an OpenPose input to an OpenSim result.
