### 2022-08-16 #4362 @danasolav
Hi @davidpagnon,

I have a few comments regarding the paper. Please address them, and then we can proceed to acceptance:

1. According to the [OpenPose](https://github.com/CMU-Perceptual-Computing-Lab/openpose) page, they allow for single-person 3D pose reconstruction and estimation, and not only 2D. Could you please explain in the paper, how your solution differs from their 3D reconstruction and why it is necessary?
2. Line 26: Do you have a reference to the statement that markers hinder natural movement?
3. Line 33: The XYZ positions of what? the joint centers?
4. Can you elaborate more on the camera calibration procedure? Which calibration method is used (with references)?
5. Can you clarify what "Tracking of the person viewed by the most cameras" means? What happens if one person is viewed by most cameras in some time frames of the video and another in other time frames?
6. Figure 1: shouldn't the 2.i camera calibration sub-figure show the calibration process (e.g., with checkerboard plates)?
7. Could you please refer to other open-source markerless 3D reconstruction tools, if relevant (e.g., pose3d, freemocap, opencap)?
8. Can you elaborate on the keypoints that need to be tracked in order to successfully reconstruct an OpenSim model, and provide more details on the interpolation and filtering processes?
9. Line 71: it is not clear what you mean by "these setup files can be taken as is". After OpenSim performs the scaling and calculation of joint angles, what are the output results? What about the errors associated with this step (e.g., if the distance between the knee and ankle keypoints changes that is an obvious error, for example)?
10. Line 80: delete "even".
11. Line 86: replace "won't" with "will not".
12. In the bulleted lists, please replace the commas with semicolons and do not capitalize the first word, or alternatively place a period at the end and capitalize the first word.
13. The options described in Pose2Simcore should be better explained in the main text. For example, it might be unclear for the reader what is the connection between OpenPose, AlphaPose, and DeepLabCut. Similarly with regard to the calibration.
14. Some of the Pose2Sim utilities have no context in the text, which makes the list hard to understand. For example, where do you get a c3d file from? what is matplotlib? It would be helpful if you included a diagram of the possible workflows with the file types, so that a reader who has not yet used the software can better understand these lists of tools and features. Otherwise, more explanation is needed when describing these tools.
15. The way you refer to file type is not consistent throughout the text (sometimes it's in parentheses, sometimes with a . before the file type, capitalized or not). For example: (json), (h5), .toml, .c3d, trc, .trc, TRC. Please use a consistent method for file types/extensions.
16. Please refer in the text to the OpenCV tools you used in your software and reference them, before the acknowledgments.

Thanks,
Dana
