### 2022-08-17 #4362 @davidpagnon
Thank you!
Version number is v2.0.1, and here is the Zenodo DOI: [![DOI](https://zenodo.org/badge/501642916.svg)](https://zenodo.org/badge/latestdoi/501642916)

I will address your points sometime this week!
