+++ Content/paper.bib
@@ -22,6 +22,17 @@
   publisher={IEEE}
 }
 
+@article{Colyer_2018,
+  title={A review of the evolution of vision-based motion analysis and the integration of advanced computer vision methods towards developing a markerless system},
+  author={Colyer, Steffi L and Evans, Murray and Cosker, Darren P and Salo, Aki IT},
+  journal={Sports medicine-open},
+  volume={4},
+  number={1},
+  pages={1--15},
+  year={2018},
+  publisher={SpringerOpen}
+}
+
 @article{Delp_2007,
   title={OpenSim: open-source software to create and analyze dynamic simulations of movement},
   author={Delp, Scott L and Anderson, Frank C and Arnold, Allison S and Loan, Peter and Habib, Ayman and John, Chand T and Guendelman, Eran and Thelen, Darryl G},
@@ -42,6 +53,35 @@
   year={2017},
   URL = {https://ieeexplore.ieee.org/document/8237518},
   DOI = {10.1109/ICCV.2017.256}
+}
+
+@article{Hartley_1997,
+  title={Triangulation},
+  author={Hartley, Richard I and Sturm, Peter},
+  journal={Computer vision and image understanding},
+  volume={68},
+  number={2},
+  pages={146--157},
+  year={1997},
+  publisher={Elsevier}
+}
+
+@misc{Hidalgo_2019,
+  author = {Hidalgo, Ginés},
+  title = {OpenPose Experimental Models},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/CMU-Perceptual-Computing-Lab/openpose_train/tree/master/experimental_models#body_25b-model---option-2-recommended}
+}
+
+@misc{Hidalgo_2021,
+  author = {Hidalgo, Ginés},
+  title = {OpenPose 3D reconstruction module},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/CMU-Perceptual-Computing-Lab/openpose/blob/master/doc/advanced/3d_reconstruction_module.md}
 }
 
 @article{Kanko_2021,
@@ -79,6 +119,17 @@
   year={2018},
   URL = {https://www.nature.com/articles/s41593-018-0209-y},
   DOI = {10.1038/s41593-018-0209-y},
+  publisher={Nature Publishing Group}
+}
+
+@article{Needham_2021,
+  title={The accuracy of several pose estimation methods for 3D joint centre localisation},
+  author={Needham, Laurie and Evans, Murray and Cosker, Darren P and Wade, Logan and McGuigan, Polly M and Bilzon, James L and Colyer, Steffi L},
+  journal={Scientific reports},
+  volume={11},
+  number={1},
+  pages={1--11},
+  year={2021},
   publisher={Nature Publishing Group}
 }
 
@@ -132,6 +183,17 @@
   URL = {https://doi.org/10.1016/j.gaitpost.2007.07.007},
   DOI = {10.1016/j.gaitpost.2007.07.007},
   publisher={Elsevier}
+}
+
+@article{Zhang_2000,
+  title={A flexible new technique for camera calibration},
+  author={Zhang, Zhengyou},
+  journal={IEEE Transactions on pattern analysis and machine intelligence},
+  volume={22},
+  number={11},
+  pages={1330--1334},
+  year={2000},
+  publisher={IEEE}
 }
 
 @article{Zheng_2022,
