+++ Content/paper.bib
@@ -1,3 +1,14 @@
+@article{Beaucage_2019,
+  title={Validation of an OpenSim full-body model with detailed lumbar spine for estimating lower lumbar spine loads during symmetric and asymmetric lifting tasks},
+  author={Beaucage-Gauvreau, Erica and Robertson, William SP and Brandon, Scott CE and Fraser, Robert and Freeman, Brian JC and Graham, Ryan B and Thewlis, Dominic and Jones, Claire F},
+  journal={Computer methods in biomechanics and biomedical engineering},
+  volume={22},
+  number={5},
+  pages={451--464},
+  year={2019},
+  publisher={Taylor \& Francis}
+}
+
 @article{Bradski_2000,
     author = {Bradski, G.},
     citeulike-article-id = {2236121},
@@ -7,6 +18,16 @@
     priority = {4},
     title = {{The OpenCV Library}},
     year = {2000}
+}
+
+@article{Butterworth_1930,
+  title={On the theory of filter amplifiers},
+  author={Butterworth, Stephen and others},
+  journal={Wireless Engineer},
+  volume={7},
+  number={6},
+  pages={536--541},
+  year={1930}
 }
 
 @article{Cao_2019,
@@ -20,6 +41,16 @@
   URL = {https://arxiv.org/abs/1611.08050},
   DOI = {10.1109/TPAMI.2019.2929257},
   publisher={IEEE}
+}
+
+@article{Cleveland_1981,
+  title={LOWESS: A program for smoothing scatterplots by robust locally weighted regression},
+  author={Cleveland, William S},
+  journal={American Statistician},
+  volume={35},
+  number={1},
+  pages={54},
+  year={1981}
 }
 
 @article{Colyer_2018,
@@ -122,6 +153,17 @@
   publisher={Nature Publishing Group}
 }
 
+ @misc{Matthis_2022, 
+  type={Python},
+  title={FreeMoCap: A free, open source markerless motion capture system},
+  rights={AGPL-3.0}, 
+  url={https://github.com/freemocap/freemocap}, 
+  author={Matthis, Jonathan Samir and Cherian, Aaron}, 
+  year={2022}, 
+  publisher = {GitHub},
+  journal = {GitHub repository},
+}
+
 @article{Needham_2021,
   title={The accuracy of several pose estimation methods for 3D joint centre localisation},
   author={Needham, Laurie and Evans, Murray and Cosker, Darren P and Wade, Logan and McGuigan, Polly M and Bilzon, James L and Colyer, Steffi L},
@@ -157,19 +199,53 @@
   publisher={Multidisciplinary Digital Publishing Institute}
 }
 
+@article{Rajagopal_2016,
+  title={Full-body musculoskeletal model for muscle-driven simulation of human gait},
+  author={Rajagopal, Apoorva and Dembia, Christopher L and DeMers, Matthew S and Delp, Denny D and Hicks, Jennifer L and Delp, Scott L},
+  journal={IEEE transactions on biomedical engineering},
+  volume={63},
+  number={10},
+  pages={2068--2079},
+  year={2016},
+  publisher={IEEE}
+}
+
 @article{Seth_2018,
-    doi = {10.1371/journal.pcbi.1006223},
-    author = {Seth, Ajay AND Hicks, Jennifer L. AND Uchida, Thomas K. AND Habib, Ayman AND Dembia, Christopher L. AND Dunne, James J. AND Ong, Carmichael F. AND DeMers, Matthew S. AND Rajagopal, Apoorva AND Millard, Matthew AND Hamner, Samuel R. AND Arnold, Edith M. AND Yong, Jennifer R. AND Lakshmikanth, Shrinidhi K. AND Sherman, Michael A. AND Ku, Joy P. AND Delp, Scott L.},
[…]
