### 2022-08-24 #4362 @davidpagnon
Thank you Dana, you raised very good points, and you may be right that the internal functioning may need to be more detailed. I elaborated much more on the "Pose2Sim core" section, which is now called "Pose2Sim method details".

> * According to the [OpenPose](https://github.com/CMU-Perceptual-Computing-Lab/openpose) page, they allow for single-person 3D pose reconstruction and estimation, and not only 2D. Could you please explain in the paper, how your solution differs from their 3D reconstruction and why it is necessary?

This is a good question, thanks for asking! It is now broached in the statement of need:
"These coordinates can then be triangulated in order to produce 3D positions. Aside from `Pose2Sim`, a number a tools are available in this regard: the experimental `OpenPose 3D reconstruction module` [@Hidalgo_2021], the `FreeMoCap` Python and Blender toolbox [@Matthis_2022], or the `pose3d` Matlab toolbox [@Sheshadri_2020]. Yet, when it comes to the biomechanical analysis of human motion, it is often more useful to obtain joint angles than joint center positions in space."

and in the Triangulation section:
"`Pose2Sim` triangulation is robust, largely because instead of using classic Direct Linear Transform (DLT) [@Hartley_1997], we propose a weighted DLT, i.e., a triangulation procedure where each OpenPose keypoint coordinate is weighted with its confidence score [@Pagnon_2021]. Other parameters can be specified, such as: [..]"

> * Line 26: Do you have a reference to the statement that markers hinder natural movement

[Colyer, Sports Med - Open, 2018] is now cited.

> * Line 33: The XYZ positions of what? the joint centers?

Replaced with "joint center positions in space"

> * Can you elaborate more on the camera calibration procedure? Which calibration method is used (with references)?

This has been detailed in the Calibration section: "The user can indicate whether cameras are going to be calibrated with a checkerboard, or if a preexisting calibration file such as provided by a Qualisys system will simply be converted.

If checkerboard calibration is chosen, the number of corners and the size of the squares have to be specified. In this case, the operator needs to take at least 10 pictures or one video per camera of the checkerboard, covering as much as the field of view as possible, with different orientations. Corners are then detected and refined with OpenCV. Detected corners can optionally be displayed for verification. Each camera is finally calibrated using OpenCV with an algorithm based on [@Zhang_2000]. The user can choose the index of the image which they want to be used as a reference for calculating extrinsic parameters."

> * Can you clarify what "Tracking of the person viewed by the most cameras" means? What happens if one person is viewed by most cameras in some time frames of the video and another in other time frames?

It has been changed to "Tracking the person of interest", and further explained in the "Tracking" section. Pose2Sim does not currently support multi-person triangulation, although it's in the to-do-list.

"One needs to differentiate the people in the background from the actual subject. The tracking step examines all possible triangulations of a chosen keypoint among all detected persons, and reprojects them on the image planes. The triangulation with the smallest reprojection error is considered to be the one associating the right person from all cameras. If the reprojection error is above a predefined threshold, the process is repeated after taking off one, or several cameras. This happens, for example, if the person of interest has exited the field of a camera, while another person is still in the background. We recommend choosing the neck point or one of the hip points. Indeed, in most cases they are the least likely to move out of the camera views."

> * Figure 1: shouldn't the 2.i camera calibration sub-figure show the calibration process (e.g., with checkerboard plates)?

I chose to represent calibration by displaying the cameras views correctly situated in the 3D scene, rather than by showing a checkerboard procedure. Indeed, here calibration can either be done with a checkerboard, or simply by converting a preexisting calibrated file. For instance, the Qualisys system makes it possible to calibrate video cameras with a classic wand procedure.

> * Could you please refer to other open-source markerless 3D reconstruction tools, if relevant (e.g., pose3d, freemocap, opencap)?

Added them in the statement of need:
"These coordinates can then be triangulated in order to produce 3D positions. Aside from `Pose2Sim`, a number a tools are available to do so: the experimental `OpenPose 3D reconstruction module` [@Hidalgo_2021], the `FreeMoCap` Python and Blender toolbox [@Matthis_2022], or the `pose3d` Matlab toolbox [@Sheshadri_2020]. Yet, when it comes to the biomechanical analysis [...]"

"three software applications are worth mentioning. [...] `OpenCap` [@Ulrich_2022] has recently been released, and offers a user friendly web application working with low-cost hardware. It predicts the coordinates of 43 anatomical markers from 20 triangulated keypoints, and imports them in `OpenSim`. However, the source code has not been released yet."

> * Can you elaborate on the keypoints that need to be tracked in order to successfully reconstruct an OpenSim model, and provide more details on the interpolation and filtering processes?

This is now explicited in the "2D keypoint detection" section: "Only 21 of the 25 keypoints detected by the default OpenPose models are tracked, since eye and ear keypoints would be redundant in the determination of the head orientation."

Interpolation and filtering processes have been detailed in the "Triangulation" and "Filtering" sections.
"Once all frames are triangulated, the ones with missing keypoint coordinates are interpolated. The interpolation method can also be chosen among linear, slinear, quadratic, and cubic."

"Different filters can be chosen, and their parameters can be adjusted. The user can choose among a zero-phase low-pass Butterworth filter [@Butterworth_1930] that they can apply either on keypoint positions or on their speeds, a LOESS filter [@Cleveland_1981], a Gaussian filter, or a median filter. Waveforms before and after filtering can be displayed and compared."

> * Line 71: it is not clear what you mean by "these setup files can be taken as is". After OpenSim performs the scaling and calculation of joint angles, what are the output results? What about the errors associated with this step (e.g., if the distance between the knee and ankle keypoints changes that is an obvious error, for example)?

I am not sure I understand this point. The output results are given as a .mot file, which is basically a table listing the time-series of the joint angles. OpenSim comes with a biomechanically constrained model, which is more than just joint limits. Bone dimensions are fixed, which means that in your example, the knee flexion/extension angle will be optimized as regards to keypoints positions, but the bone length will not be affected. This is why model-based inverse kinematics is particularly robust, as compared to geometry-based kinematics. In any case, keypoints may occasionally be mislabelled, but this also happens in marker-based capture, in particular due to skin artifact movements.

Maybe you thought I was referring to result files instead of setup ones? This sentence has been deleted, and the OpenSim procedure has been detailed in the next section.

"OpenSim allows for much more accurate and robust results [@Pagnon_2022], since it constrains kinematics to an individually scaled and physically accurate skeletal model. [...] Unlike in marker-based capture, and despite the aforementioned systematic errors, keypoints detection hardly depends on the subject, the operator, nor the context. For this reason, the scaling and the inverse kinematic steps are straight-forward, and the provided setup files require little to no adjusting."

> * Line 80: delete "even".

Done.

> * Line 86: replace "won't" with "will not".

Done.

> * In the bulleted lists, please replace the commas with semicolons and do not capitalize the first word, or alternatively place a period at the end and capitalize the first word.

Done.

> * The options described in Pose2Sim core should be better explained in the main text. For example, it might be unclear for the reader what is the connection between OpenPose, AlphaPose, and DeepLabCut. Similarly with regard to the calibration.

As quoted ahead, this has been completely rewritten.

> * Some of the Pose2Sim utilities have no context in the text, which makes the list hard to understand. For example, where do you get a c3d file from? what is matplotlib? It would be helpful if you included a diagram of the possible workflows with the file types, so that a reader who has not yet used the software can better understand these lists of tools and features. Otherwise, more explanation is needed when describing these tools.

I trimmed this part a lot, integrated it in the previous section, and added a diagram for clarity.

> * The way you refer to file type is not consistent throughout the text (sometimes it's in parentheses, sometimes with a . before the file type, capitalized or not). For example: (json), (h5), .toml, .c3d, trc, .trc, TRC. Please use a consistent method for file types/extensions.

Corrected.

> * Please refer in the text to the OpenCV tools you used in your software and reference them, before the acknowledgments.

Done.
