+++ Content/paper.bib
@@ -2,6 +2,7 @@
   title={Validation of an OpenSim full-body model with detailed lumbar spine for estimating lower lumbar spine loads during symmetric and asymmetric lifting tasks},
   author={Beaucage-Gauvreau, Erica and Robertson, William SP and Brandon, Scott CE and Fraser, Robert and Freeman, Brian JC and Graham, Ryan B and Thewlis, Dominic and Jones, Claire F},
   journal={Computer methods in biomechanics and biomedical engineering},
+  DOI = {10.1080/10255842.2018.1564819},
   volume={22},
   number={5},
   pages={451--464},
@@ -10,14 +11,14 @@
 }
 
 @article{Bradski_2000,
-    author = {Bradski, G.},
-    citeulike-article-id = {2236121},
-    journal = {Dr. Dobb's Journal of Software Tools},
-    keywords = {bibtex-import},
-    posted-at = {2008-01-15 19:21:54},
-    priority = {4},
-    title = {{The OpenCV Library}},
-    year = {2000}
+  author = {Bradski, G.},
+  citeulike-article-id = {2236121},
+  journal = {Dr. Dobb's Journal of Software Tools},
+  keywords = {bibtex-import},
+  posted-at = {2008-01-15 19:21:54},
+  priority = {4},
+  title = {The OpenCV Library},
+  year = {2000}
 }
 
 @article{Butterworth_1930,
@@ -46,6 +47,7 @@
 @article{Cleveland_1981,
   title={LOWESS: A program for smoothing scatterplots by robust locally weighted regression},
   author={Cleveland, William S},
+  DOI={10.2307/2683591},
   journal={American Statistician},
   volume={35},
   number={1},
@@ -57,6 +59,7 @@
   title={A review of the evolution of vision-based motion analysis and the integration of advanced computer vision methods towards developing a markerless system},
   author={Colyer, Steffi L and Evans, Murray and Cosker, Darren P and Salo, Aki IT},
   journal={Sports medicine-open},
+  DOI={10.1186/s40798-018-0139-y},
   volume={4},
   number={1},
   pages={1--15},
@@ -90,6 +93,7 @@
   title={Triangulation},
   author={Hartley, Richard I and Sturm, Peter},
   journal={Computer vision and image understanding},
+  DOI={10.1006/cviu.1997.0547},
   volume={68},
   number={2},
   pages={146--157},
@@ -168,6 +172,7 @@
   title={The accuracy of several pose estimation methods for 3D joint centre localisation},
   author={Needham, Laurie and Evans, Murray and Cosker, Darren P and Wade, Logan and McGuigan, Polly M and Bilzon, James L and Colyer, Steffi L},
   journal={Scientific reports},
+  DOI={10.1038/s41598-021-00212-x},
   volume={11},
   number={1},
   pages={1--11},
@@ -203,6 +208,7 @@
   title={Full-body musculoskeletal model for muscle-driven simulation of human gait},
   author={Rajagopal, Apoorva and Dembia, Christopher L and DeMers, Matthew S and Delp, Denny D and Hicks, Jennifer L and Delp, Scott L},
   journal={IEEE transactions on biomedical engineering},
+  DOI={10.1109/tbme.2016.2586891},
   volume={63},
   number={10},
   pages={2068--2079},
@@ -211,7 +217,7 @@
 }
 
 @article{Seth_2018,
-  doi = {10.1371/journal.pcbi.1006223},
[…]
