### 2022-08-29 d2e0bfe
Periods at the end of lists
