### 2022-09-02 77ea4e8
Shortly elaborate on bone and joint constraints
