### 2022-09-08 #4362 @danasolav
@davidpagnon
Please consider the following comments and then issue a new tagged release of the software and archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other). Then, please post the version number and archive DOI here.
Line 66: please delete "pieces of".
Line 133: consider replacing with "If there are not enough cameras left" for improved clarity.
After that, I will proceed with acceptance.
