+++ Content/paper.bib
@@ -23,7 +23,7 @@
 
 @article{Butterworth_1930,
   title={On the theory of filter amplifiers},
-  author={Butterworth, Stephen and others},
+  author={Butterworth, Stephen},
   journal={Wireless Engineer},
   volume={7},
   number={6},
