### 2022-09-09 #4362 @davidpagnon
Some of the DOIs are missing simply because they don't have any. For example, there is none for The OpenCV library, despite I followed [their recommendation](https://github.com/opencv/opencv/wiki/CiteOpenCV). Some others, such as the paper from Butterworth, are just too old to have a DOI. The last ones are GitHub repositories.
I suppose it is not part of what you want me to change?

Aside from that, I edited the two lines you suggested accordingly, and issued a new tagged release.
Pose2Sim v0.2.2
Zenodo DOI always resolving to the last version: [10.5281/zenodo.6998625](https://doi.org/10.5281/zenodo.6998625)
Current Zenodo DOI: [10.5281/zenodo.7064026](https://doi.org/10.5281/zenodo.7064026)
