### 2022-09-12 #4362 @danielskatz
I've suggested some changes in https://github.com/perfanalytics/pose2sim/pull/10, mostly to use markdown rather than forcing formatting.
