### 2022-09-13 #4362 @davidpagnon
Thanks @danielskatz! I suggested one last rearrangement, moving some text above the picture to avoid large white spaces.
