### 2022-09-13 #4362 @danielskatz
@davidpagnon - note that the list on page 2 now has elements 2.1, 2.2, etc, which no longer matches Figure 1's caption or the text in the figure itself.  How do you think this should be addressed?
