### 2022-09-13 #4362 @davidpagnon
You mean, the 2.1 vs. 2.i ? I'm not sure how to do it in markdown and I need to leave now, I'll find a way tomorrow! I do notice a typo in the figure too, "cameras calibration" instead of "camera calibration". I'll correct it, too.
