### 2022-09-13 #4362 @danielskatz
yes, correct.  I don't know how to change it in markdown either, but you could change the figure caption and text in the figure to use 2.1 etc, which might be easier.
